requirements.txt
pytest
coverage
python-dotenv
watchdog
blinker
Flask
flake8
biopython
ratelimit
backoff